fix: redid the rendering of cards and corrected the filling of the cards
