+++ src/pages/main-search/mainSearch.ts
@@ -1,15 +1,16 @@
-import products from '../../products.json'
-import Search, { Product } from '../../core/search';
+import products from '../../products.json';
+import MainPage, { Product } from '../main-page/main';
 
-class SearchProducts extends Search {
+class SearchProducts extends MainPage {
   private searchBar: HTMLElement;
 
   constructor () {
-    super();
+    super('search');
     this.searchBar = document.getElementById('searchBar') as HTMLInputElement;
   }
 
   searchProduct (): void {
+    document.getElementById('searchBar')?.addEventListener('focus', (event) => { const target = event.target as HTMLSelectElement; target.value = ''; this.render(products, 'search') }); // для очистки инпута, когда нажимаем снова (чтобы не висело постоянно)
     this.searchBar.addEventListener('keyup', (event) => {
       const target = event.target as HTMLSelectElement;
       const searchString = target.value.toLowerCase();
@@ -17,18 +18,18 @@ class SearchProducts extends Search {
       const resultRender = products.filter((product) => {
         return product.title.toLowerCase().includes(searchString);
       })
-      this.render(resultRender);
+      this.renderSearch(resultRender);
     })
   }
 
-  render (searchResult: Product[]): void {
+  renderSearch (searchResult: Product[]): void {
     const productContainer = document.querySelector('.products__container') as HTMLElement;
     productContainer.innerHTML = '';
     const arr: Product[] = [];
-    searchResult.forEach((item, index) => {
+    searchResult.forEach((item) => {
       arr.push(item);
     })
-    this.renderElements(arr);
+    this.render(arr, 'search');
   }
 }
